+++ CHANGELOG.rst
@@ -33,6 +33,10 @@ Changed
 - Extra keyword arguments passed to ``Spectrum.plot`` are now forwarded to ``Axes.plot``.
 - Tools (e.g., ``Cube``, ``Maps``) can now be accessed from the ``marvin`` namespace (e.g., ``marvin.tools.Cube`` or ``marvin.tools.cube.Cube``).
 
+Refactored
+^^^^^^^^^^
+- Moved `marvin.core.core` to `marvin.tools.core` and split the mixins into `marvin.tools.mixins`.
+
 
 [2.2.5] - 2018/04/26
 --------------------